Client side tests and restructuring
Tests are moved into app/test and server/test with test/ handling tests
for both.
