+++ app/test/example_spec.js
@@ -1,0 +1,7 @@
+import { expect } from 'chai';
+
+describe('example tests', () => {
+  it('should always be true', () => {
+    expect(true).to.equal(true);
+  });
+});
